requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
transformers>=4.38.0
torch>=2.2.0
Pillow>=10.2.0
